refactor: 게임 시도 횟수 입력 부분 작은 함수로 분리
