+++ src/main/java/racingcar/InputHandler.java
@@ -48,5 +48,26 @@ public class InputHandler {
         return cars;
     }
 
+    public static int inputTryCount() {
+        int tryCount = readTryCount();
+        validatTryCount(tryCount);
+        return tryCount;
+    }
 
+    private static int readTryCount() {
+        try {
+            System.out.println("시도할 횟수는 몇 회인가요?");
+            return Integer.parseInt(Console.readLine());
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("시도할 횟수는 숫자로 입력해야 합니다.");
+
+        }
+    }
+
+    private static void validatTryCount(int count) {
+        if (count <= 0) {
+            throw new IllegalArgumentException("시도할 횟수는 1번 이상이어야 합니다.");
+        }
+        System.out.println();
+    }
 }
